stat_window: Set xticks to integer
Our dataset only allows for full integer-range numbers, which needs to
be represented by the axis itself. Thus, set it to just integer, which
also limits in reusability in the code itself for further
customizations(and enhanced stats), but allows also for better sight on
the dataset.

TODO is to actually externalize that as well, because it's getting more
and more spaghettie.