LiGi-Timers: "Normal" Timers are decreasing, not increasing.....
Makes also development a lot easier. Currently we have to keep track of
different values in case we want to start and stop, which is now gone.
Essentially, i also could just delete the `goal`, because it isn't
really needed for anything more then special little features like
in-class-resets of times.