LiGi-Timer: Enable Literals, use steady_clock, remove test client
Got the CRTP-Client finally moved out, also using the steady_clock. It's
essential we do, because the high resolution one is just too much. We
dont need it, simply.

The Literals-Namespace is going to survive in more areas of this current
library.
It's just a toy-library to reimplement patterns and little programs,
sure. But it's actually helpfull to me.

Im thinking of a provider-class autospawning a child into a thread. Not
exactly sure if i should even do that...