LiGi-timer: State-Prep for stopped-by-mechanic
Currently, we can't know wether the timer returned becasue we stopped it
through the global stopper variable or any other mechanism, like just
running out of time left.

This could be a potential idea, but also could be just awful. I could
implement that in the App itself, ofc, because i've got the control over
the stop variable.