BasicallyHTTP: A simple self written http server
Currently, just using C to serve. I plan to, as you can see on the top,
encapsulate as much as i can and try to reach somehow a fun little toy
project with it.

Im not sure if thats a good idea, so this could be abandoned in a few
days already. But for now it seems fun!